Add a wrapper for the Tumblr API
Tumblr sometimes updates the names of their API calls, which causes things to break.
Once this wrapper is implemented in the rest of the code, any changes in the API that break the code should be fixable by updating the wrapper instead of combing through the project for calls to functions whose names have changed.